refactor(mobile): formatPrice를 shared/utils로 분리 및 로케일 명시 (#236)
- Intl.NumberFormat 로케일을 undefined → 'ko-KR'로 명시하여 기기 간 일관성 보장
- formatPrice를 subscription 도메인에서 shared/utils/format.ts로 이동 (SRP)
- 관련 테스트를 format.test.ts로 분리
