+++ apps/mobile/src/features/subscription/presentations/components/SubscriptionPlanCard.tsx
@@ -2,10 +2,10 @@ import { HStack } from '@src/shared/ui/HStack/HStack';
 import { Text } from '@src/shared/ui/Text/Text';
 import { VStack } from '@src/shared/ui/VStack/VStack';
 import { cn } from '@src/shared/utils/cn';
+import { formatPrice } from '@src/shared/utils/format';
 import { Card, PressableFeedback } from 'heroui-native';
 import { View } from 'react-native';
 import {
-  formatPrice,
   getAnnualDiscountPercent,
   getMonthlyEquivalent,
   type SubscriptionPlan,
